# swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "paths": {
        "/allStore": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "get_all_store",
                "tags": [
                    "default"
                ]
            }
        },
        "/hello": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "get_hello_world",
                "tags": [
                    "default"
                ]
            }
        },
        "/login": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "post_login",
                "tags": [
                    "default"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "get_login",
                "tags": [
                    "default"
                ]
            }
        },
        "/myStore/{userID}": {
            "parameters": [
                {
                    "name": "userID",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "get_my_store",
                "tags": [
                    "default"
                ]
            }
        },
        "/myStore/{userID}/read/{postID}": {
            "parameters": [
                {
                    "name": "userID",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                },
                {
                    "name": "postID",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "get_my_store_read",
                "tags": [
                    "default"
                ]
            }
        },
        "/register": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "post_register",
                "tags": [
                    "default"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "get_register",
                "tags": [
                    "default"
                ]
            }
        },
        "/register/quipuCheck": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "post_check",
                "tags": [
                    "default"
                ]
            }
        },
        "/store/{userID}": {
            "parameters": [
                {
                    "name": "userID",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "get_store",
                "tags": [
                    "default"
                ]
            }
        },
        "/store/{userID}/write/{type}": {
            "parameters": [
                {
                    "name": "userID",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                },
                {
                    "name": "type",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "operationId": "post_store_write",
                "tags": [
                    "default"
                ]
            }
        }
    },
    "info": {
        "title": "API",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "tags": [
        {
            "name": "default",
            "description": "Default namespace"
        }
    ],
    "responses": {
        "ParseError": {
            "description": "When a mask can't be parsed"
        },
        "MaskError": {
            "description": "When any error occurs on mask"
        }
    },
    "host": "localhost:8080"
}